feat(miniapp): enhance miniapp types and update imports for consistency
